move specs to specs folder
