+++ pkg/executor/specs/scraperspec.go
@@ -1,0 +1,78 @@
+package specs
+
+import (
+	"bytes"
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/kubeshop/testkube/pkg/executor"
+	"github.com/kubeshop/testkube/pkg/executor/options"
+	"github.com/kubeshop/testkube/pkg/utils"
+	"go.uber.org/zap"
+	batchv1 "k8s.io/api/batch/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/yaml"
+	"sigs.k8s.io/kustomize/kyaml/yaml/merge2"
+
+	kyaml "sigs.k8s.io/kustomize/kyaml/yaml"
+)
+
+// NewScraperJobSpec is a method to create new scraper job spec
+func NewScraperJobSpec(log *zap.SugaredLogger, options options.JobOptions) (*batchv1.Job, error) {
+	tmpl, err := utils.NewTemplate("job").Parse(options.ScraperTemplate)
+	if err != nil {
+		return nil, fmt.Errorf("creating job spec from scraper template error: %w", err)
+	}
+
+	options.Jsn = strings.ReplaceAll(options.Jsn, "'", "''")
+	var buffer bytes.Buffer
+	if err = tmpl.ExecuteTemplate(&buffer, "job", options); err != nil {
+		return nil, fmt.Errorf("executing job spec scraper template: %w", err)
+	}
+
+	var job batchv1.Job
+	jobSpec := buffer.String()
+	if options.ScraperTemplateExtensions != "" {
+		tmplExt, err := utils.NewTemplate("jobExt").Parse(options.ScraperTemplateExtensions)
+		if err != nil {
+			return nil, fmt.Errorf("creating scraper extensions spec from executor template error: %w", err)
+		}
+
+		var bufferExt bytes.Buffer
+		if err = tmplExt.ExecuteTemplate(&bufferExt, "jobExt", options); err != nil {
+			return nil, fmt.Errorf("executing scraper extensions spec executor template: %w", err)
+		}
+
+		if jobSpec, err = merge2.MergeStrings(bufferExt.String(), jobSpec, false, kyaml.MergeOptions{}); err != nil {
+			return nil, fmt.Errorf("merging scraper spec executor templates: %w", err)
+		}
+	}
+
+	log.Debug("Scraper job specification", jobSpec)
+	decoder := yaml.NewYAMLOrJSONDecoder(bytes.NewBufferString(jobSpec), len(jobSpec))
+	if err := decoder.Decode(&job); err != nil {
+		return nil, fmt.Errorf("decoding scraper job spec error: %w", err)
+	}
+
+	envs := append(executor.RunnerEnvVars, corev1.EnvVar{Name: "RUNNER_CLUSTERID", Value: options.ClusterID})
+	if options.ArtifactRequest != nil && options.ArtifactRequest.StorageBucket != "" {
+		envs = append(envs, corev1.EnvVar{Name: "RUNNER_BUCKET", Value: options.ArtifactRequest.StorageBucket})
+	} else {
+		envs = append(envs, corev1.EnvVar{Name: "RUNNER_BUCKET", Value: os.Getenv("STORAGE_BUCKET")})
+	}
+
+	if options.HTTPProxy != "" {
+		envs = append(envs, corev1.EnvVar{Name: "HTTP_PROXY", Value: options.HTTPProxy})
+	}
+
+	if options.HTTPSProxy != "" {
+		envs = append(envs, corev1.EnvVar{Name: "HTTPS_PROXY", Value: options.HTTPSProxy})
+	}
+
+	for i := range job.Spec.Template.Spec.Containers {
+		job.Spec.Template.Spec.Containers[i].Env = append(job.Spec.Template.Spec.Containers[i].Env, envs...)
+	}
+
+	return &job, nil
+}
